Reorder update passes
Split out anim pass.
This is strictly cut-and-paste, no code has been changed.
