+++ masonry/src/passes/update.rs
@@ -77,206 +77,109 @@ fn run_single_update_pass(
     }
 }
 
-// --- MARK: UPDATE POINTER ---
-pub(crate) fn run_update_pointer_pass(root: &mut RenderRoot) {
-    let pointer_pos = root.last_mouse_pos.map(|pos| (pos.x, pos.y).into());
+// --- MARK: UPDATE TREE ---
+fn update_widget_tree(
+    global_state: &mut RenderRootState,
+    mut widget: ArenaMut<'_, Box<dyn Widget>>,
+    mut state: ArenaMut<'_, WidgetState>,
+) {
+    let _span = widget.item.make_trace_span().entered();
+    let id = state.item.id;
 
-    // Release pointer capture if target can no longer hold it.
-    if let Some(id) = root.state.pointer_capture_target {
-        if !root.is_still_interactive(id) {
-            root.state.pointer_capture_target = None;
-            run_on_pointer_event_pass(root, &PointerEvent::new_pointer_leave());
-        }
+    if !state.item.children_changed {
+        return;
     }
+    state.item.children_changed = false;
 
-    // -- UPDATE HOVERED WIDGETS --
-    let mut next_hovered_widget = if let Some(pos) = pointer_pos {
-        // TODO - Apply scale?
-        root.get_root_widget()
-            .find_widget_at_pos(pos)
-            .map(|widget| widget.id())
-    } else {
-        None
-    };
-    // If the pointer is captured, it can either hover its capture target or nothing.
-    if let Some(capture_target) = root.state.pointer_capture_target {
-        if next_hovered_widget != Some(capture_target) {
-            next_hovered_widget = None;
-        }
-    }
+    {
+        let mut ctx = RegisterCtx {
+            widget_state_children: state.children.reborrow_mut(),
+            widget_children: widget.children.reborrow_mut(),
+            #[cfg(debug_assertions)]
+            registered_ids: Vec::new(),
+        };
+        // The widget will call `RegisterCtx::register_child` on all its children,
+        // which will add the new widgets to the arena.
+        widget.item.register_children(&mut ctx);
 
-    // "Hovered path" means the widget which is considered hovered, and all its parents.
-    let prev_hovered_path = std::mem::take(&mut root.state.hovered_path);
-    let next_hovered_path = get_id_path(root, next_hovered_widget);
-
-    let mut hovered_set = HashSet::new();
-    for widget_id in &next_hovered_path {
-        hovered_set.insert(*widget_id);
-    }
-
-    trace!("prev_hovered_path: {:?}", prev_hovered_path);
-    trace!("next_hovered_path: {:?}", next_hovered_path);
-
-    // This algorithm is written to be resilient to future changes like reparenting and multiple
-    // cursors. In theory it's O(Depth² * CursorCount) in the worst case, which isn't too bad
-    // (cursor count is usually 1 or 2, depth is usually small), but in practice it's virtually
-    // always O(Depth * CursorCount) because we only need to update the hovered status of the
-    // widgets that changed.
-    // The above assumes that accessing the widget tree is O(1) for simplicity.
-
-    fn update_hovered_status_of(
-        root: &mut RenderRoot,
-        widget_id: WidgetId,
-        hovered_set: &HashSet<WidgetId>,
-    ) {
-        run_targeted_update_pass(root, Some(widget_id), |widget, ctx| {
-            let is_hovered = hovered_set.contains(&ctx.widget_id());
-
-            if ctx.widget_state.is_hovered != is_hovered {
-                widget.on_status_change(ctx, &StatusChange::HoveredChanged(is_hovered));
+        #[cfg(debug_assertions)]
+        {
+            let children_ids = widget.item.children_ids();
+            for child_id in ctx.registered_ids {
+                if !children_ids.contains(&child_id) {
+                    panic!(
+                        "Error in '{}' #{}: method register_children() called \
+                        RegisterCtx::register_child() on child #{}, which isn't \
+                        in the list returned by children_ids()",
+                        widget.item.short_type_name(),
+                        id.to_raw(),
+                        child_id.to_raw()
+                    );
+                }
             }
-            ctx.widget_state.is_hovered = is_hovered;
-        });
-    }
-
-    // TODO - Make sure widgets are iterated from the bottom up.
-    // TODO - Document the iteration order for update_pointer pass.
-    for widget_id in prev_hovered_path.iter().copied() {
-        if root.widget_arena.has(widget_id)
-            && root.widget_arena.get_state_mut(widget_id).item.is_hovered
-                != hovered_set.contains(&widget_id)
-        {
-            update_hovered_status_of(root, widget_id, &hovered_set);
         }
-    }
-    for widget_id in next_hovered_path.iter().copied() {
-        if root.widget_arena.has(widget_id)
-            && root.widget_arena.get_state_mut(widget_id).item.is_hovered
-                != hovered_set.contains(&widget_id)
-        {
-            update_hovered_status_of(root, widget_id, &hovered_set);
+
+        #[cfg(debug_assertions)]
+        for child_id in widget.item.children_ids() {
+            if widget.children.get_child(child_id.to_raw()).is_none() {
+                panic!(
+                    "Error in '{}' #{}: method register_children() did not call \
+                    RegisterCtx::register_child() on child #{} returned by children_ids()",
+                    widget.item.short_type_name(),
+                    id.to_raw(),
+                    child_id.to_raw()
+                );
+            }
         }
     }
 
-    // -- UPDATE CURSOR --
-
-    // If the pointer is captured, its cursor always reflects the
-    // capture target, even when not hovered.
-    let cursor_source = root.state.pointer_capture_target.or(next_hovered_widget);
-
-    let new_cursor = if let Some(cursor_source) = cursor_source {
-        let (widget, state) = root.widget_arena.get_pair(cursor_source);
-        state.item.cursor.unwrap_or(widget.item.get_cursor())
-    } else {
-        CursorIcon::Default
-    };
-
-    if root.state.cursor_icon != new_cursor {
-        root.state
-            .emit_signal(RenderRootSignal::SetCursor(new_cursor));
+    if state.item.is_new {
+        let mut ctx = LifeCycleCtx {
+            global_state,
+            widget_state: state.item,
+            widget_state_children: state.children.reborrow_mut(),
+            widget_children: widget.children.reborrow_mut(),
+        };
+        widget.item.lifecycle(&mut ctx, &LifeCycle::WidgetAdded);
+        trace!(
+            "{} received LifeCycle::WidgetAdded",
+            widget.item.short_type_name()
+        );
+        state.item.accepts_pointer_interaction = widget.item.accepts_pointer_interaction();
+        state.item.accepts_focus = widget.item.accepts_focus();
+        state.item.accepts_text_input = widget.item.accepts_text_input();
+        state.item.is_new = false;
     }
 
-    root.state.cursor_icon = new_cursor;
-    root.state.hovered_path = next_hovered_path;
+    // We can recurse on this widget's children, because they have already been added
+    // to the arena above.
+    let parent_state = state.item;
+    recurse_on_children(
+        id,
+        widget.reborrow_mut(),
+        state.children,
+        |widget, mut state| {
+            update_widget_tree(global_state, widget, state.reborrow_mut());
+            parent_state.merge_up(state.item);
+        },
+    );
 }
 
-// ----------------
+pub(crate) fn run_update_widget_tree_pass(root: &mut RenderRoot) {
+    let _span = info_span!("update_new_widgets").entered();
 
-// --- MARK: UPDATE FOCUS ---
-pub(crate) fn run_update_focus_pass(root: &mut RenderRoot) {
-    // If the focused widget is disabled, stashed or removed, we set
-    // the focused id to None
-    if let Some(id) = root.state.next_focused_widget {
-        if !root.is_still_interactive(id) {
-            root.state.next_focused_widget = None;
-        }
-    }
-
-    let prev_focused = root.state.focused_widget;
-    let next_focused = root.state.next_focused_widget;
-
-    // "Focused path" means the focused widget, and all its parents.
-    let prev_focused_path = std::mem::take(&mut root.state.focused_path);
-    let next_focused_path = get_id_path(root, next_focused);
-
-    let mut focused_set = HashSet::new();
-    for widget_id in &next_focused_path {
-        focused_set.insert(*widget_id);
-    }
-
-    trace!("prev_focused_path: {:?}", prev_focused_path);
-    trace!("next_focused_path: {:?}", next_focused_path);
-
-    // This is the same algorithm as the one in
-    // run_update_pointer_pass
-    // See comment in that function.
-
-    fn update_focused_status_of(
-        root: &mut RenderRoot,
-        widget_id: WidgetId,
-        focused_set: &HashSet<WidgetId>,
-    ) {
-        run_targeted_update_pass(root, Some(widget_id), |widget, ctx| {
-            let has_focus = focused_set.contains(&ctx.widget_id());
-
-            if ctx.widget_state.has_focus != has_focus {
-                widget.on_status_change(ctx, &StatusChange::ChildFocusChanged(has_focus));
-            }
-            ctx.widget_state.has_focus = has_focus;
-        });
-    }
-
-    // TODO - Make sure widgets are iterated from the bottom up.
-    // TODO - Document the iteration order for update_focus pass.
-    for widget_id in prev_focused_path.iter().copied() {
-        if root.widget_arena.has(widget_id)
-            && root.widget_arena.get_state_mut(widget_id).item.has_focus
-                != focused_set.contains(&widget_id)
-        {
-            update_focused_status_of(root, widget_id, &focused_set);
-        }
-    }
-    for widget_id in next_focused_path.iter().copied() {
-        if root.widget_arena.has(widget_id)
-            && root.widget_arena.get_state_mut(widget_id).item.has_focus
-                != focused_set.contains(&widget_id)
-        {
-            update_focused_status_of(root, widget_id, &focused_set);
-        }
-    }
-
-    if prev_focused != next_focused {
-        let was_ime_active = root.state.is_ime_active;
-        let is_ime_active = if let Some(id) = next_focused {
-            root.widget_arena.get_state(id).item.accepts_text_input
-        } else {
-            false
+    if root.root.incomplete() {
+        let mut ctx = RegisterCtx {
+            widget_state_children: root.widget_arena.widget_states.root_token_mut(),
+            widget_children: root.widget_arena.widgets.root_token_mut(),
+            #[cfg(debug_assertions)]
+            registered_ids: Vec::new(),
         };
-        root.state.is_ime_active = is_ime_active;
-
-        run_single_update_pass(root, prev_focused, |widget, ctx| {
-            widget.on_status_change(ctx, &StatusChange::FocusChanged(false));
-        });
-        run_single_update_pass(root, next_focused, |widget, ctx| {
-            widget.on_status_change(ctx, &StatusChange::FocusChanged(true));
-        });
-
-        if prev_focused.is_some() && was_ime_active {
-            root.state.emit_signal(RenderRootSignal::EndIme);
-        }
-        if next_focused.is_some() && is_ime_active {
-            root.state.emit_signal(RenderRootSignal::StartIme);
-        }
-
-        if let Some(id) = next_focused {
-            let ime_area = root.widget_arena.get_state(id).item.get_ime_area();
-            root.state
-                .emit_signal(RenderRootSignal::new_ime_moved_signal(ime_area));
-        }
+        ctx.register_child(&mut root.root);
     }
 
-    root.state.focused_widget = root.state.next_focused_widget;
-    root.state.focused_path = next_focused_path;
+    let (root_widget, mut root_state) = root.widget_arena.get_pair_mut(root.root.id());
+    update_widget_tree(&mut root.state, root_widget, root_state.reborrow_mut());
 }
 
 // ----------------
@@ -401,200 +304,6 @@ pub(crate) fn run_update_stashed_pass(root: &mut RenderRoot) {
 
 // ----------------
 
-// --- MARK: UPDATE SCROLL ---
-// This pass will update scroll positions in cases where a widget has requested to be
-// scrolled into view (usually a textbox getting text events).
-// Each parent that implements scrolling will update its scroll position to ensure the
-// child is visible. (If the target area is larger than the parent, the parent will try
-// to show the top left of that area.)
-pub(crate) fn run_update_scroll_pass(root: &mut RenderRoot) {
-    let _span = info_span!("update_scroll").entered();
-
-    let scroll_request_targets = std::mem::take(&mut root.state.scroll_request_targets);
-    for (target, rect) in scroll_request_targets {
-        let mut target_rect = rect;
-
-        run_targeted_update_pass(root, Some(target), |widget, ctx| {
-            let event = LifeCycle::RequestPanToChild(rect);
-            widget.lifecycle(ctx, &event);
-
-            // TODO - We should run the compose method after this, so
-            // translations are updated and the rect passed to parents
-            // is more accurate.
-
-            let state = &ctx.widget_state;
-            target_rect = target_rect + state.translation + state.origin.to_vec2();
-        });
-    }
-}
-
-// ----------------
-
-// --- MARK: UPDATE ANIM ---
-fn update_anim_for_widget(
-    global_state: &mut RenderRootState,
-    mut widget: ArenaMut<'_, Box<dyn Widget>>,
-    mut state: ArenaMut<'_, WidgetState>,
-    elapsed_ns: u64,
-) {
-    let _span = widget.item.make_trace_span().entered();
-
-    if !state.item.needs_anim {
-        return;
-    }
-    state.item.needs_anim = false;
-
-    // Most passes reset their `needs` and `request` flags after the call to
-    // the widget method, but it's valid and expected for `request_anim` to be
-    // set in response to `AnimFrame`.
-    if state.item.request_anim {
-        state.item.request_anim = false;
-        let mut ctx = LifeCycleCtx {
-            global_state,
-            widget_state: state.item,
-            widget_state_children: state.children.reborrow_mut(),
-            widget_children: widget.children.reborrow_mut(),
-        };
-        widget
-            .item
-            .lifecycle(&mut ctx, &LifeCycle::AnimFrame(elapsed_ns));
-    }
-
-    let id = state.item.id;
-    let parent_state = state.item;
-    recurse_on_children(
-        id,
-        widget.reborrow_mut(),
-        state.children,
-        |widget, mut state| {
-            update_anim_for_widget(global_state, widget, state.reborrow_mut(), elapsed_ns);
-            parent_state.merge_up(state.item);
-        },
-    );
-}
-
-/// Run the animation pass.
-pub(crate) fn run_update_anim_pass(root: &mut RenderRoot, elapsed_ns: u64) {
-    let _span = info_span!("update_anim").entered();
-
-    let (root_widget, mut root_state) = root.widget_arena.get_pair_mut(root.root.id());
-    update_anim_for_widget(
-        &mut root.state,
-        root_widget,
-        root_state.reborrow_mut(),
-        elapsed_ns,
-    );
-}
-
-// ----------------
-
-// --- MARK: UPDATE TREE ---
-fn update_widget_tree(
-    global_state: &mut RenderRootState,
-    mut widget: ArenaMut<'_, Box<dyn Widget>>,
-    mut state: ArenaMut<'_, WidgetState>,
-) {
-    let _span = widget.item.make_trace_span().entered();
-    let id = state.item.id;
-
-    if !state.item.children_changed {
-        return;
-    }
-    state.item.children_changed = false;
-
-    {
-        let mut ctx = RegisterCtx {
-            widget_state_children: state.children.reborrow_mut(),
-            widget_children: widget.children.reborrow_mut(),
-            #[cfg(debug_assertions)]
-            registered_ids: Vec::new(),
-        };
-        // The widget will call `RegisterCtx::register_child` on all its children,
-        // which will add the new widgets to the arena.
-        widget.item.register_children(&mut ctx);
-
-        #[cfg(debug_assertions)]
-        {
-            let children_ids = widget.item.children_ids();
-            for child_id in ctx.registered_ids {
-                if !children_ids.contains(&child_id) {
-                    panic!(
-                        "Error in '{}' #{}: method register_children() called \
-                        RegisterCtx::register_child() on child #{}, which isn't \
-                        in the list returned by children_ids()",
-                        widget.item.short_type_name(),
-                        id.to_raw(),
-                        child_id.to_raw()
-                    );
-                }
-            }
-        }
-
-        #[cfg(debug_assertions)]
-        for child_id in widget.item.children_ids() {
-            if widget.children.get_child(child_id.to_raw()).is_none() {
-                panic!(
-                    "Error in '{}' #{}: method register_children() did not call \
-                    RegisterCtx::register_child() on child #{} returned by children_ids()",
-                    widget.item.short_type_name(),
-                    id.to_raw(),
-                    child_id.to_raw()
-                );
-            }
-        }
-    }
-
-    if state.item.is_new {
-        let mut ctx = LifeCycleCtx {
-            global_state,
-            widget_state: state.item,
-            widget_state_children: state.children.reborrow_mut(),
-            widget_children: widget.children.reborrow_mut(),
-        };
-        widget.item.lifecycle(&mut ctx, &LifeCycle::WidgetAdded);
-        trace!(
-            "{} received LifeCycle::WidgetAdded",
-            widget.item.short_type_name()
-        );
-        state.item.accepts_pointer_interaction = widget.item.accepts_pointer_interaction();
-        state.item.accepts_focus = widget.item.accepts_focus();
-        state.item.accepts_text_input = widget.item.accepts_text_input();
-        state.item.is_new = false;
-    }
-
-    // We can recurse on this widget's children, because they have already been added
-    // to the arena above.
-    let parent_state = state.item;
-    recurse_on_children(
-        id,
-        widget.reborrow_mut(),
-        state.children,
-        |widget, mut state| {
-            update_widget_tree(global_state, widget, state.reborrow_mut());
-            parent_state.merge_up(state.item);
-        },
-    );
-}
-
-pub(crate) fn run_update_widget_tree_pass(root: &mut RenderRoot) {
-    let _span = info_span!("update_new_widgets").entered();
-
-    if root.root.incomplete() {
-        let mut ctx = RegisterCtx {
-            widget_state_children: root.widget_arena.widget_states.root_token_mut(),
-            widget_children: root.widget_arena.widgets.root_token_mut(),
-            #[cfg(debug_assertions)]
-            registered_ids: Vec::new(),
-        };
-        ctx.register_child(&mut root.root);
-    }
-
-    let (root_widget, mut root_state) = root.widget_arena.get_pair_mut(root.root.id());
-    update_widget_tree(&mut root.state, root_widget, root_state.reborrow_mut());
-}
-
-// ----------------
-
 // --- MARK: UPDATE FOCUS CHAIN ---
 
 // TODO https://github.com/linebender/xilem/issues/376 - Some implicit invariants:
@@ -669,4 +378,237 @@ pub(crate) fn run_update_focus_chain_pass(root: &mut RenderRoot) {
         root_state.reborrow_mut(),
         &mut dummy_focus_chain,
     );
+}
+
+// ----------------
+
+// --- MARK: UPDATE FOCUS ---
+pub(crate) fn run_update_focus_pass(root: &mut RenderRoot) {
+    // If the focused widget is disabled, stashed or removed, we set
+    // the focused id to None
+    if let Some(id) = root.state.next_focused_widget {
+        if !root.is_still_interactive(id) {
+            root.state.next_focused_widget = None;
+        }
+    }
+
+    let prev_focused = root.state.focused_widget;
+    let next_focused = root.state.next_focused_widget;
+
+    // "Focused path" means the focused widget, and all its parents.
+    let prev_focused_path = std::mem::take(&mut root.state.focused_path);
+    let next_focused_path = get_id_path(root, next_focused);
+
+    let mut focused_set = HashSet::new();
+    for widget_id in &next_focused_path {
+        focused_set.insert(*widget_id);
+    }
+
+    trace!("prev_focused_path: {:?}", prev_focused_path);
+    trace!("next_focused_path: {:?}", next_focused_path);
+
+    // This is the same algorithm as the one in
+    // run_update_pointer_pass
+    // See comment in that function.
+
+    fn update_focused_status_of(
+        root: &mut RenderRoot,
+        widget_id: WidgetId,
+        focused_set: &HashSet<WidgetId>,
+    ) {
+        run_targeted_update_pass(root, Some(widget_id), |widget, ctx| {
+            let has_focus = focused_set.contains(&ctx.widget_id());
+
+            if ctx.widget_state.has_focus != has_focus {
+                widget.on_status_change(ctx, &StatusChange::ChildFocusChanged(has_focus));
+            }
+            ctx.widget_state.has_focus = has_focus;
+        });
+    }
+
+    // TODO - Make sure widgets are iterated from the bottom up.
+    // TODO - Document the iteration order for update_focus pass.
+    for widget_id in prev_focused_path.iter().copied() {
+        if root.widget_arena.has(widget_id)
+            && root.widget_arena.get_state_mut(widget_id).item.has_focus
+                != focused_set.contains(&widget_id)
+        {
+            update_focused_status_of(root, widget_id, &focused_set);
+        }
+    }
+    for widget_id in next_focused_path.iter().copied() {
+        if root.widget_arena.has(widget_id)
+            && root.widget_arena.get_state_mut(widget_id).item.has_focus
+                != focused_set.contains(&widget_id)
+        {
+            update_focused_status_of(root, widget_id, &focused_set);
+        }
+    }
+
+    if prev_focused != next_focused {
+        let was_ime_active = root.state.is_ime_active;
+        let is_ime_active = if let Some(id) = next_focused {
+            root.widget_arena.get_state(id).item.accepts_text_input
+        } else {
+            false
+        };
+        root.state.is_ime_active = is_ime_active;
+
+        run_single_update_pass(root, prev_focused, |widget, ctx| {
+            widget.on_status_change(ctx, &StatusChange::FocusChanged(false));
+        });
+        run_single_update_pass(root, next_focused, |widget, ctx| {
+            widget.on_status_change(ctx, &StatusChange::FocusChanged(true));
+        });
+
+        if prev_focused.is_some() && was_ime_active {
+            root.state.emit_signal(RenderRootSignal::EndIme);
+        }
+        if next_focused.is_some() && is_ime_active {
+            root.state.emit_signal(RenderRootSignal::StartIme);
+        }
+
+        if let Some(id) = next_focused {
+            let ime_area = root.widget_arena.get_state(id).item.get_ime_area();
+            root.state
+                .emit_signal(RenderRootSignal::new_ime_moved_signal(ime_area));
+        }
+    }
+
+    root.state.focused_widget = root.state.next_focused_widget;
+    root.state.focused_path = next_focused_path;
+}
+
+// ----------------
+
+// --- MARK: UPDATE SCROLL ---
+// This pass will update scroll positions in cases where a widget has requested to be
+// scrolled into view (usually a textbox getting text events).
+// Each parent that implements scrolling will update its scroll position to ensure the
+// child is visible. (If the target area is larger than the parent, the parent will try
+// to show the top left of that area.)
+pub(crate) fn run_update_scroll_pass(root: &mut RenderRoot) {
+    let _span = info_span!("update_scroll").entered();
+
+    let scroll_request_targets = std::mem::take(&mut root.state.scroll_request_targets);
+    for (target, rect) in scroll_request_targets {
+        let mut target_rect = rect;
+
+        run_targeted_update_pass(root, Some(target), |widget, ctx| {
+            let event = LifeCycle::RequestPanToChild(rect);
+            widget.lifecycle(ctx, &event);
+
+            // TODO - We should run the compose method after this, so
+            // translations are updated and the rect passed to parents
+            // is more accurate.
+
+            let state = &ctx.widget_state;
+            target_rect = target_rect + state.translation + state.origin.to_vec2();
+        });
+    }
+}
+
+// ----------------
+
+// --- MARK: UPDATE POINTER ---
+pub(crate) fn run_update_pointer_pass(root: &mut RenderRoot) {
+    let pointer_pos = root.last_mouse_pos.map(|pos| (pos.x, pos.y).into());
+
+    // Release pointer capture if target can no longer hold it.
+    if let Some(id) = root.state.pointer_capture_target {
+        if !root.is_still_interactive(id) {
+            root.state.pointer_capture_target = None;
+            run_on_pointer_event_pass(root, &PointerEvent::new_pointer_leave());
+        }
+    }
+
+    // -- UPDATE HOVERED WIDGETS --
+    let mut next_hovered_widget = if let Some(pos) = pointer_pos {
+        // TODO - Apply scale?
+        root.get_root_widget()
+            .find_widget_at_pos(pos)
+            .map(|widget| widget.id())
+    } else {
+        None
+    };
+    // If the pointer is captured, it can either hover its capture target or nothing.
+    if let Some(capture_target) = root.state.pointer_capture_target {
+        if next_hovered_widget != Some(capture_target) {
+            next_hovered_widget = None;
+        }
+    }
+
+    // "Hovered path" means the widget which is considered hovered, and all its parents.
+    let prev_hovered_path = std::mem::take(&mut root.state.hovered_path);
+    let next_hovered_path = get_id_path(root, next_hovered_widget);
+
+    let mut hovered_set = HashSet::new();
+    for widget_id in &next_hovered_path {
+        hovered_set.insert(*widget_id);
+    }
+
+    trace!("prev_hovered_path: {:?}", prev_hovered_path);
+    trace!("next_hovered_path: {:?}", next_hovered_path);
+
+    // This algorithm is written to be resilient to future changes like reparenting and multiple
+    // cursors. In theory it's O(Depth² * CursorCount) in the worst case, which isn't too bad
+    // (cursor count is usually 1 or 2, depth is usually small), but in practice it's virtually
+    // always O(Depth * CursorCount) because we only need to update the hovered status of the
+    // widgets that changed.
+    // The above assumes that accessing the widget tree is O(1) for simplicity.
+
+    fn update_hovered_status_of(
+        root: &mut RenderRoot,
+        widget_id: WidgetId,
+        hovered_set: &HashSet<WidgetId>,
+    ) {
+        run_targeted_update_pass(root, Some(widget_id), |widget, ctx| {
+            let is_hovered = hovered_set.contains(&ctx.widget_id());
+
+            if ctx.widget_state.is_hovered != is_hovered {
+                widget.on_status_change(ctx, &StatusChange::HoveredChanged(is_hovered));
+            }
+            ctx.widget_state.is_hovered = is_hovered;
+        });
+    }
+
+    // TODO - Make sure widgets are iterated from the bottom up.
+    // TODO - Document the iteration order for update_pointer pass.
+    for widget_id in prev_hovered_path.iter().copied() {
+        if root.widget_arena.has(widget_id)
+            && root.widget_arena.get_state_mut(widget_id).item.is_hovered
+                != hovered_set.contains(&widget_id)
+        {
+            update_hovered_status_of(root, widget_id, &hovered_set);
+        }
+    }
+    for widget_id in next_hovered_path.iter().copied() {
+        if root.widget_arena.has(widget_id)
+            && root.widget_arena.get_state_mut(widget_id).item.is_hovered
+                != hovered_set.contains(&widget_id)
+        {
+            update_hovered_status_of(root, widget_id, &hovered_set);
+        }
+    }
+
+    // -- UPDATE CURSOR --
+
+    // If the pointer is captured, its cursor always reflects the
+    // capture target, even when not hovered.
+    let cursor_source = root.state.pointer_capture_target.or(next_hovered_widget);
+
+    let new_cursor = if let Some(cursor_source) = cursor_source {
+        let (widget, state) = root.widget_arena.get_pair(cursor_source);
+        state.item.cursor.unwrap_or(widget.item.get_cursor())
+    } else {
+        CursorIcon::Default
+    };
+
+    if root.state.cursor_icon != new_cursor {
+        root.state
+            .emit_signal(RenderRootSignal::SetCursor(new_cursor));
+    }
+
+    root.state.cursor_icon = new_cursor;
+    root.state.hovered_path = next_hovered_path;
 }
